Boostrapped frontend

+++ backend/api/categorie.js
@@ -1,10 +1,11 @@
 const Categorie = require('../models/categorie');
 const Article = require('../models/article');
+const User = require('../models/user');
+const VerifyJWT = require('../middleware/verifyJWT');
 
 const router = require('express').Router();
 
 router.get('/', async (req, res) => {
-    console.log("/categorie");
     try {
         const categories = await Categorie.find().sort({ name: 1 });
         res.json(categories);
@@ -13,9 +14,10 @@ router.get('/', async (req, res) => {
     }
 });
 
-router.delete('/del/:id', async (req, res) => {
-    console.log("/categorie/del");
+router.delete('/del/:id', VerifyJWT, async (req, res) => {
     if (!req.params.id) return res.status(400).send('ID is required');
+    const user = await User.findById(req.user.id);
+    if (!['admin'].includes(user.role)) return res.status(403).send('You do not have permission to delete a category');
     try {
         await Categorie.findByIdAndDelete(req.params.id);
         await Article.updateMany({ categories: req.params.id }, { $pull: { categories: req.params.id } });
@@ -25,10 +27,10 @@ router.delete('/del/:id', async (req, res) => {
     }
 });
 
-
-router.post('/add', async (req, res) => {
-    console.log("/categorie/add");
+router.post('/add', VerifyJWT, async (req, res) => {
     if (!req.body.name) return res.status(400).send('Name is required');
+    const user = await User.findById(req.user.id);
+    if (!['admin'].includes(user.role)) return res.status(403).send('You do not have permission to delete a category');
     try {
         await Categorie.create({ name: req.body.name });
         res.json({ message: 'Category successfully added' });
@@ -38,4 +40,3 @@ router.post('/add', async (req, res) => {
 });
 
 module.exports = router;
-
